исправяет замечания проверяющего
